feat: expand opcode decoders, pcapng capture with client isolation, and outbound decryption (#2)
## Summary

Ten commits since `main` that grow the decoder set, make live capture
follow the client's own connections, and add outbound (client→server)
decryption and richer DebugUi inspection.

### Decoders
- Read opcodes as full 32-bit `uint` (TW carries a category byte in the
upper 16 bits; the old `ushort` collided `0x00021208` with `0x1208`).
Opcodes now display as 8-digit hex everywhere.
- Expand from 24 to ~90 opcode decoders across pet, world/region,
stance, motion, props, URLs, guild, equipment, combat targeting, buff KV
and movement. Replay decode coverage on real captures rises from ~26% to
~90%.
- Decode the owner-only `0x5209` character snapshot (600–700 elems)
structurally, with C-parameter sections, fixtures and alignment tests.
Fully defensive: malformed input yields a partial record.
- Surface `PartyWindowUpdate` fields; rename `ConditionUpdate2` →
`CharacterConditionUpdate` and the `ChannelCharacterInfo` model →
`CharacterData` to match decoded shapes.

### Capture
- Record live sessions as pcapng (nanosecond timestamps, interface
metadata) via a background frame recorder, replacing the classic pcap
writer.
- Capture filter follows the game process's own connections so a channel
switch is never missed; a client-traffic filter drops frames from other
processes sharing the host address, upstream of both recorder and
pipeline.
- Design doc for client traffic isolation and channel switching.

### Network
- Classify frame direction by which end is the game server, so outbound
packets are labelled and kept.
- Optional client→server decryption using Mabinogi's client cipher,
keyed by the 4-byte connection key the server sends at connection open.
Exposed in DebugUi as a "Capture out" toggle, off by default.

### DebugUi
- Collapsible decoded-payload tree, hex pane with elem-range
highlighting (right-click an elem to jump to its bytes).
- Per-packet export from the packet list: single-frame pcapng, full JSON
envelope, compact specimen JSON. Works on multi-selection.

## Test plan
- [x] `dotnet build -p:TreatWarningsAsErrors=true` — 0 warnings, 0
errors
- [x] `dotnet test` (full suite incl. Npcap-dependent tests) — 528
passed, 3 skipped, 0 failed
- [x] CI green on this PR (after two follow-up commits: a CS0067 pragma
on a test stub, and a trait-based Npcap test filter replacing the stale
class-name filter)



---------

+++ src/Mabipacade.Core/Model/MabiPacket.cs
@@ -3,7 +3,7 @@ namespace Mabipacade.Core.Model;
 public sealed record MabiPacket(
     DateTime TimestampUtc,
     Direction Direction,
-    ushort Op,
+    uint Op,
     ulong EntityId,
     IReadOnlyList<MessageElem> Elems,
     object? Decoded)
