+++ src/Mabipacade.Core/Plugins/IPacketDecoder.cs
@@ -2,6 +2,6 @@ namespace Mabipacade.Core.Plugins;
 
 public interface IPacketDecoder
 {
-    ushort Op { get; }
+    uint Op { get; }
     object Decode(DecoderInput input);
 }
